Added basic Item class <- GameObject, RandCell + glyphs
RandCell returns the x,y image.Point of a random cell which is not a wall.
Careful, as in the future, we'll probably want to construct this to take into
consideration other obstacles besides walls which are not walkable cells.
In other words, instead of opt out strategy of IsWall, IsWater, etc, do IsWalkable.
Consider using collision function in game/map.go